refactor: update auth routes and restructure form handling
- Update signin and signup services
- Remove deprecated form interfaces and auth types
- Add new models directory
- Update core generated files
