+++ chama-frontend/src/data/user-role.ts
@@ -1,0 +1,4 @@
+export enum UserRole {
+    ADMIN,
+    MEMBER
+}
