--- chama-frontend/src/types/auth.ts
@@ -1,10 +1,0 @@
-export type UserRole = "admin" | "member"
-
-export type UserProfile = {
-    id: string
-    firstName: string
-    lastName: string
-    email: string;
-    phoneNumber: string;
-    role: UserRole
-}
